// ui/src/generated/ScaleKind.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type ScaleKind = "Chromatic" | "Major" | "Minor" | "HarmonicMinor" | "MelodicMinor" | "MajorPentatonic" | "MinorPentatonic" | "Blues" | "Dorian" | "Phrygian" | "Lydian" | "Mixolydian" | "Locrian" | "WholeTone";